refactor(logistics): move cache adapter to infrastructure
