+++ server/features/logistics/application/README.md
@@ -174,14 +174,17 @@ intactos: el guard global los **subsume**, no los reemplaza.
 `release`/`start`/`complete` (resto del lifecycle) permanecen fuera de la capa y
 siguen usando el default del helper de lifecycle compartido.
 GET/POST/location/time-windows de field visits permanecen fuera de M09 y se
-difieren al thin-route M15. **M12** ya está implementado (pendiente de merge): movió
+difieren al thin-route M15. **M12** (mergeado en PR #1509) movió
 `db-logistics.ts` completo a `infrastructure/db-logistics.ts` con las transacciones
 intactas, dejando un shim en el root. Es **infraestructura de persistencia** y **no
 cambia esta capa**: `application` sigue sin importar `db-*` — el guard de M11 lo
 prohíbe **por nombre de módulo**, de modo que sigue siendo correcto antes y después
-del move. El siguiente milestone es **M13** (cache adapter para
-`server/lib/logistics-route-plans-cache.ts`). Cada puerto nuevo se introduce junto
-con su primer consumidor real — nunca como interfaz vacía anticipada.
+del move. **M13** movió el cache de route plans a
+`infrastructure/logistics-route-plans-cache.ts` (shim en `server/lib`), también sin
+tocar esta capa y **sin puerto de cache anticipado**: el puerto se definirá junto
+con su primer consumidor real cuando la ruta se adelgace en **M14**, el siguiente
+milestone. Cada puerto nuevo se introduce junto con su primer consumidor real —
+nunca como interfaz vacía anticipada.
 
 ## Qué NO hacer
