+++ server/features/logistics/infrastructure/README.md
@@ -1,8 +1,9 @@
 # Logistics · infrastructure (implementación de puertos)
 
 > Capa **infrastructure** del contexto Logistics. **Contiene código** desde M02b
-> (adaptador de DB para el breach de SLA) y, desde **M12 — implementado /
-> pendiente de merge**, la **persistencia canónica** del contexto.
+> (adaptador de DB para el breach de SLA); desde **M12 (mergeado en PR #1509)**,
+> la **persistencia canónica** del contexto; y desde **M13**, el **cache de
+> route plans**.
 > Ver la frontera del contexto en [`../README.md`](../README.md) y el contrato en
 > [ARCH-2](../../../../docs/architecture/backend-boundary-adr.md).
 
@@ -38,8 +39,17 @@ de persistencia y el I/O real del contexto.
   (`./db-logistics.ts`) en vez del shim raíz, conservando el import dinámico (lazy)
   de `markOverdueActiveClinicSlaInstancesBreached`, la delegación al núcleo de
   dominio y las signatures.
+- **`logistics-route-plans-cache.ts`** (M13) — **cache canónico** de route plans:
+  move **byte-idéntico** desde `server/lib/logistics-route-plans-cache.ts`
+  (107 LOC, **cero imports**, 9 exports). In-memory puro (dos `Map` module-level),
+  TTL de 5 minutos, expiración lazy, miss = `null`, invalidación completa y por
+  prefijo `clinic:`/`clinic:+plan:`. La **construcción de claves** y el header
+  `X-Logistics-Cache` (HIT/MISS) **no viven aquí**: pertenecen a
+  `server/routes/logistics-route-plans.fastify.ts` hasta M14. Sin puerto de
+  cache anticipado. La pureza (cero imports) y el shim quedan fijados por el
+  guard de frontera.
 
-## Shim de compatibilidad en el root
+## Shims de compatibilidad fuera de la capa
 
 `server/db-logistics.ts` queda como **shim temporal**: sólo re-exporta la superficie
 pública desde `./features/logistics/infrastructure/db-logistics.ts`. No importa
@@ -49,17 +59,24 @@ transacciones, y no declara default export. Existe porque las rutas legacy
 (estática y dinámicamente) hasta **M14–M16**; al cerrarse esa secuencia, el shim
 desaparece.
 
+`server/lib/logistics-route-plans-cache.ts` es, desde M13, el **shim del cache**:
+un único `export *` hacia el canónico de esta capa, sin lógica ni exports propios.
+Existe porque `server/routes/logistics-route-plans.fastify.ts` sigue importándolo
+hasta **M14**; al adelgazarse esa ruta, el shim desaparece.
+
 **Sin cambios de schema ni de migraciones**: M12 no toca `drizzle/schema.ts`,
 `drizzle/**`, `migrations/**`, endpoints ni contratos HTTP.
 
 ## Qué vivirá aquí (futuro, no ahora)
 
-- **M13 — cache adapter** para `server/lib/logistics-route-plans-cache.ts` (TTL/keys
-  preservados; contract-test de runtime existente). **Pendiente**: M12 no lo adelanta.
+- Nada nuevo planificado para esta capa: los siguientes milestones de la Fase C
+  (**M14–M16**, thin routes, y **M17**, cierre) adelgazan `server/routes/**` y
+  retiran los shims; no añaden módulos aquí.
 
 ## Qué NO hacer
 
-No mover la cache todavía (M13). No crear adaptadores vacíos. No tocar
+No crear adaptadores vacíos. No introducir un puerto de cache anticipado (se
+define junto con su primer consumidor real en M14). No tocar
 `drizzle/schema.ts` ni migraciones. No reparticionar las transacciones del archivo
 canónico. No importar `application`, `routes` ni Fastify desde esta capa (el guard
 de frontera lo bloquea).
